requirements.txt
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
deepdiff==8.6.0
gitdb==4.0.12
GitPython==3.1.41
idna==3.10
mypy_extensions==1.1.0
orderly-set==5.5.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
smmap==5.0.2
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
